requirements.txt
Pillow>=11.3.0